refact: Simplify code with new Foundation class
